cleanup pt1
