+++ app/src/app/components/contents/manageAccess/updatePermission/RemovePermissionsForUpdate.tsx
@@ -26,47 +26,49 @@ export const RemovePermissionsForUpdate = ({ removePermission, rolePermissions }
   return (
     <>
       <Label>Permissions To Remove</Label>
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>
-          <Button variant="outline" role="combobox" aria-expanded={open} className="w-[350px] justify-between">
-            {value ? constructFoundRolePermission(value) : "Select Permission..."}
-            <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent className="w-[350px] p-0">
-          <Command>
-            <CommandInput placeholder="Search framework..." />
-            <CommandList>
-              <CommandEmpty>No permissions found.</CommandEmpty>
-              <CommandGroup>
-                {rolePermissions.map((rolePermission) => (
-                  <CommandItem
-                    key={rolePermission.id}
-                    value={rolePermission.id.toString()}
-                    onSelect={(currentValue) => {
-                      setValue(currentValue === value?.id.toString() ? undefined : rolePermission);
-                      setOpen(false);
-                    }}
-                  >
-                    <Check
-                      className={cn("mr-2 h-4 w-4", value?.id === rolePermission.id ? "opacity-100" : "opacity-0")}
-                    />
-                    {constructPermissionName(rolePermission)}
-                  </CommandItem>
-                ))}
-              </CommandGroup>
-            </CommandList>
-          </Command>
-        </PopoverContent>
-      </Popover>
-      <Button
-        variant="outline"
-        size="icon"
-        disabled={value?.permission === undefined}
-        onClick={() => removePermission(value as RolePermission)}
-      >
-        <SquarePlus className="h-4 w-4" />
-      </Button>
+      <div className="flex space-x-3">
+        <Popover open={open} onOpenChange={setOpen}>
+          <PopoverTrigger asChild>
+            <Button variant="outline" role="combobox" aria-expanded={open} className="w-[350px] justify-between">
+              {value ? constructFoundRolePermission(value) : "Select Permission..."}
+              <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+            </Button>
+          </PopoverTrigger>
+          <PopoverContent className="w-[350px] p-0">
+            <Command>
+              <CommandInput placeholder="Search permission..." />
+              <CommandList>
+                <CommandEmpty>No permissions found.</CommandEmpty>
+                <CommandGroup>
+                  {rolePermissions.map((rolePermission) => (
+                    <CommandItem
+                      key={rolePermission.id}
+                      value={rolePermission.id.toString()}
+                      onSelect={(currentValue) => {
+                        setValue(currentValue === value?.id.toString() ? undefined : rolePermission);
+                        setOpen(false);
+                      }}
+                    >
+                      <Check
+                        className={cn("mr-2 h-4 w-4", value?.id === rolePermission.id ? "opacity-100" : "opacity-0")}
+                      />
+                      {constructPermissionName(rolePermission)}
+                    </CommandItem>
+                  ))}
+                </CommandGroup>
+              </CommandList>
+            </Command>
+          </PopoverContent>
+        </Popover>
+        <Button
+          variant="outline"
+          size="icon"
+          disabled={value?.permission === undefined}
+          onClick={() => removePermission(value as RolePermission)}
+        >
+          <SquarePlus className="h-4 w-4" />
+        </Button>
+      </div>
     </>
   );
 };
